Searialization bug in QueryItem fixed.
QueryItem has an attribute logger which is being searialized when
searializing QueryItem object resulting in exception. The cosmosdb
spring dependency need to be updated to latest version to fix this. But
the version which has the fix is a major release and no backward
compatiability exists. This will result in lot of changes. For now we
are going with a mixin approach to solve this issue

Related github issue:
https://github.com/microsoft/spring-data-cosmosdb/issues/423